Fix incorrect window length calculation

+++ helpers.h
@@ -18,10 +18,10 @@
 
 typedef unsigned char byte;
 
-char checksum(char *frame, size_t count);
-size_t create_frame(unsigned int seq_num, char *frame, char *data, size_t data_size, bool eot);
-void create_ack(unsigned int seq_num, char *ack, bool error);
-bool read_frame(unsigned int *seq_num, char *data, size_t *data_size, bool *eot, char *frame);
-bool read_ack(unsigned int *seq_num, bool *error, char *ack);
+char checksum(char *frame, int count);
+int create_frame(int seq_num, char *frame, char *data, int data_size, bool eot);
+void create_ack(int seq_num, char *ack, bool error);
+bool read_frame(int *seq_num, char *data, int *data_size, bool *eot, char *frame);
+bool read_ack(int *seq_num, bool *error, char *ack);
 
 #endif
